requirements.txt
django>=4.2.0,<4.2.6
djangorestframework
pyyaml
requests
django-cors-headers